refactor: domain, ui 분리.
