+++ src/main/java/racing/CarRacingApplicationMain.java
@@ -1,5 +1,9 @@
 package racing;
 
+import racing.domain.Race;
+import racing.ui.InputView;
+import racing.ui.OutputView;
+
 public class CarRacingApplicationMain {
     public static void main(String[] args) {
         Race race = InputView.createRace();
