+++ src/main/java/racing/domain/MovingRule.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public interface MovingRule {
     boolean isPossibleToMove();
