+++ src/main/java/racing/domain/RaceResult.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import java.util.ArrayList;
 import java.util.List;
