+++ src/test/java/racing/domain/RaceTest.java
@@ -1,8 +1,7 @@
-package racing;
+package racing.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
